+++ apps/internal-client/pages/index.tsx
@@ -1,57 +1,19 @@
-import { BaseConfig, getProjectsByTarget } from '@pr-helper-hub/common-react';
-import { GetStaticPropsResult } from 'next';
-import {
-  DashboardPage,
-  DashboardPageProps,
-} from '../components/dashboard-page/dashboard-page';
-import { loadNxGraph } from '../utils/load-nx-graph';
+/* eslint-disable-next-line */
+export interface IndexProps {}
 
 /**
- * The Main index landing page. Renders the dashboard page
+ * The primary landing page, or the actual app.
+ * See:
+ * core-ui for actual app implementation.
  *
- * @param props DashboardPageProps
+ * @param props the index props
  */
-export function Index(props: DashboardPageProps) {
-  return <DashboardPage {...props} />;
+export function Index(props: IndexProps) {
+  return (
+    <div>
+      <h1>Welcome to Index!</h1>
+    </div>
+  );
 }
 
 export default Index;
-
-/**
- * Returns the props for the index page.
- */
-export async function getStaticProps(): Promise<
-  GetStaticPropsResult<DashboardPageProps>
-> {
-  const config = (() => {
-    const common = {
-      sha: process.env.GITHUB_SHA ?? '',
-      ref_type: process.env.GITHUB_REF_TYPE ?? '',
-      date: new Date().toISOString(),
-    };
-    if (process.env.GITHUB_REF_TYPE === 'branch')
-      return {
-        ...common,
-        branch: process.env.GITHUB_REF ?? '',
-      };
-    // Fallback just return tag, as branch release models usually are used.
-    return {
-      ...common,
-      tag: process.env.GITHUB_REF ?? '',
-    };
-  })() as BaseConfig;
-
-  const nxGraph = await loadNxGraph();
-
-  const typedocProjects = getProjectsByTarget({
-    graph: nxGraph,
-    target: 'typedoc',
-  });
-
-  return {
-    props: {
-      config,
-      typedocProjects,
-    },
-  };
-}
